A preliminary implementation for the micropost status feed.
Although the micropost form is actually now working, users can’t
immediately see the results of a successful submission because the
current Home page doesn’t display any microposts. If you like, you can
verify that the form shown in the browser (http://localhost:3000/) is
working by submitting a valid entry and then navigating to the profile
page to see the post, but that’s rather cumbersome. It would be far
better to have a feed of microposts that includes the user’s own posts,
as mocked up in Figure 13.13.

(In Chapter 14: Following users, we’ll generalize this feed to include
the microposts of users being followed by the current user, à la
Twitter.)

Since each user should have a feed, we are led naturally to a feed
method in the User model, which will initially just select all the
microposts belonging to the current user. We’ll accomplish this using
the where method on the Micropost model (seen briefly before in Section
11.3.3.1:
https://www.railstutorial.org/book/account_activation#sec-exercises_activation_test_and_refactoring),
as shown in commit: A preliminary implementation for the micropost
status feed

The question mark in

```
Micropost.where("user_id = ?", id)
```

ensures that id is properly escaped before being included in the
underlying SQL query, thereby avoiding a serious security hole called
SQL injection. The id attribute here is just an integer (i.e., self.id,
the unique ID of the user), so there is no danger of SQL injection in
this case, but it does no harm, and always escaping variables injected
into SQL statements is a good habit to cultivate.

Alert readers might note at this point that the code in this commit is
essentially equivalent to writing

```
def feed
  microposts
end
```

We’ve used the code in this commit instead because it generalizes much
more naturally to the full status feed needed in Chapter 14: Following
users.