Adding a feed instance variable to the home action.
To use the feed in the sample application, we add an @feed_items instance variable for the current user’s (paginated) feed, as in this commit, and then add a status feed partial (commit: The status feed partial) to the Home page (commit: Adding a status feed to the Home page). Note that, now that there are two lines that need to be run when the user is logged in, in the code in this commit changes

```
@micropost = current_user.microposts.build if logged_in?
```
from commit: Adding a micropost instance variable to the home action to

```
  if logged_in?
    @micropost  = current_user.microposts.build
    @feed_items = current_user.feed.paginate(page: params[:page])
  end
```

thereby moving the conditional from the end of the line to an if-end statement.