Adding a status feed to the Home page.
We can add the feed to the Home page by rendering the feed partial as
usual (in this commit). The result is a display of the feed on the Home
page, as required (in the browser: http://localhost:3000/).

At this point, creating a new micropost works as expected, as seen in
Figure 13.15. There is one subtlety, though: on failed micropost
submission, the Home page expects an @feed_items instance variable, so
failed submissions currently break. The easiest solution is to suppress
the feed entirely by assigning it an empty array, as shown in Listing
13.50. (Unfortunately, returning a paginated feed doesn’t work in this
case. Implement it and click on a pagination link to see why.)