Adding a delete link to the micropost partial.
The last piece of functionality to add to the Microposts resource is the
ability to destroy posts. As with user deletion (Section 10.4.2: The
destroy action), we accomplish this with “delete” links, as mocked up in
Figure 13.16

(https://www.railstutorial.org/book/user_microposts#fig-micropost_delete_links_mockup).

Unlike that case, which restricted user destruction to admin users, the
delete links will work only for microposts created by the current user.

Our first step is to add a delete link to the micropost partial as in
commit: A partial for showing a single micropost. The result appears in
this commit.