The Microposts controller destroy action.
The next step is to define a destroy action in the Microposts
controller, which is analogous to the user case in commit: A before
filter restricting the destroy action to admins. The main difference is
that, rather than using an @user variable with an admin_user before
filter, we’ll find the micropost through the association, which will
automatically fail if a user tries to delete another user’s micropost.
We’ll put the resulting find inside a correct_user before filter, which
checks that the current user actually has a micropost with the given id.
The result appears in this commit.

Note that the destroy method in this commit redirects to the URL
```
request.referrer || root_url
```

This uses the request.referrer method, which is related to the
request.original_url variable used in friendly forwarding (Section
10.2.3: Friendly forwarding), and is just the previous URL (in this
case, the Home page). This is convenient because microposts appear on
both the Home page and on the user’s profile page, so by using
request.referrer we arrange to redirect back to the page issuing the
delete request in both cases. If the referring URL is nil (as is the
case inside some tests), in this commit sets the root_url as the default
using the || operator. (Compare to the default options defined in
commit: Adding a log_in_as helper.)

With the code as above, the result of destroying the second-most recent
post appears in Figure 13.17.

Exercises -

1. Create a new micropost and then delete it. What are the contents of
the DELETE command in the server log?

Parameters:
{"authenticity_token"=>"iziOjlV14cWayAic7coA8nFnB3otnQIqpSQbSoNr4pGkqvqLdrbp9+ZFpugUX5M4Ts1iE+w5BRvCjS5U13QPsA==",
"id"=>"301"}
  User Load (0.0ms)  SELECT  "users".* FROM "users" WHERE "users"."id" =
? LIMIT ?  [["id", 1], ["LIMIT", 1]]
  Micropost Load (1.0ms)  SELECT  "microposts".* FROM "microposts" WHERE
"microposts"."user_id" = ? AND "microposts"."id" = ? ORDER BY
"microposts"."created_at" DESC LIMIT ?  [["user_id", 1], ["id", 301],
["LIMIT", 1]]
   (0.0ms)  begin transaction
  SQL (4.0ms)  DELETE FROM "microposts" WHERE "microposts"."id" = ?
[["id", 301]]
   (6.0ms)  commit transaction
2. Confirm directly in the browser that the line redirect_to
request.referrer || root_url can be replaced with the line
redirect_back(fallback_location: root_url). (This method was added in
Rails 5.)