Adding a micropost with a different owner.
With the code in Section 13.3.4: Destroying microposts, the Micropost
model and interface are complete. All that’s left is writing a short
Microposts controller test to check authorization and a micropost
integration test to tie it all together.

We’ll start by adding a few microposts with different owners to the
micropost fixtures, as shown in this commit. (We’ll be using only one
for now, but we’ve put in the others for future reference.)