An integration test for the micropost interface. green
The test appears in this commit. See if you can connect the lines in commit: An integration test for the micropost interface. to the steps mentioned above.

Because we wrote working application code first, the test suite should be green:

```
$ rails test
```

Exercises -

1. For each of the four scenarios indicated by comments in Listing 13.55 (starting with “Invalid submission”), comment out application code to get the corresponding test to red, then uncomment to get back to green.

Commented out the following lines:

def create
  #@micropost = current_user.microposts.build(micropost_params)
  #if @micropost.save
  #  flash[:success] = "Micropost created!"
  #  redirect_to root_url
  #else
  #  @feed_items = []
  #  render 'static_pages/home'
  #end
end

ERROR["test_micropost_interface", MicropostsInterfaceTest, 35.51100836193655]
 test_micropost_interface#MicropostsInterfaceTest (35.51s)
NoMethodError:         NoMethodError: undefined method `document' for nil:NilClass
            test/integration/microposts_interface_test.rb:17:in `block in <class:MicropostsInterfaceTest>'

=E

def destroy
  # @micropost.destroy
  # flash[:success] = "Micropost deleted"
  # # redirect_to request.referrer || root_url
  # redirect_back(fallback_location: root_url) # added for 13.3.4  ex. 2
end

FAIL["test_micropost_interface", MicropostsInterfaceTest, 39.76235947292298]
test_micropost_interface#MicropostsInterfaceTest (39.76s)
       "Micropost.count" didn't change by -1.
       Expected: 38
         Actual: 39
       test/integration/microposts_interface_test.rb:29:in `block in <class:MicropostsInterfaceTest>'

==F

2. Add tests for the sidebar micropost count (including proper pluralization). Provided is templete that will help get you started

I'm not sure how we have 34 micropost for under micheal but I've did a search in the outputted html for "34 microposts" and found this value.

After reviewing the micropost.yml, I see that the user michael is associated with 34 entires - 30 of them created in the 30.times loop and the other 4 create individually in the fixture.