Adding CarrierWave to the Gemfile.
Now that we’ve added support for all relevant micropost actions, in this
section we’ll make it possible for microposts to include images as well
as text. We’ll start with a basic version good enough for development
use, and then add a series of enhancements to make image upload
production-ready.

Adding image upload involves two main visible elements: a form field for
uploading an image and the micropost images themselves. A mockup of the
resulting “Upload image” button and micropost photo appears in [Figure
13.18](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/micropost_image_mockup.png).

To handle an uploaded image and associate it with the Micropost model,
we’ll use the CarrierWave image uploader. To get started, we need to
include the carrierwave gem in the Gemfile (in this commit). For
completeness, Listing 13.58 also includes the mini_magick and fog-aws
gems needed for image resizing (Section 13.4.3: Image resizing) and
image upload in production (Section 13.4.4: Image upload in production),
as well as a specific version of the nokogiri gem, which is a dependency
for fog-aws. (We wouldn’t ordinarily need to specify a dependency like
nokogiri explicitly, but it turns out the dependency specification in
fog-aws allows for an earlier, insecure version of nokogiri, so we need
to take the care to update the version number by hand.)

Then we install as usual:

```
$ bundle install
```
CarrierWave adds a Rails generator for creating an image uploader, which
we’ll use to make an uploader for an image called picture:16

```
$ rails generate uploader Picture
```

Images uploaded with CarrierWave should be associated with a
corresponding attribute in an Active Record model, which simply contains
the name of the image file in a string field. The resulting augmented
data model for microposts appears in [Figure
13.19](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/micropost_model_picture.png).

```
$ rails generate migration add_picture_to_microposts picture:string
$ rails db:migrate
```

Images uploaded with CarrierWave should be associated with a
corresponding attribute in an Active Record model, which simply contains
the name of the image file in a string field. The resulting augmented
data model for microposts appears in Figure 13.19.

To add the required picture attribute to the Micropost model, we
generate a migration and migrate the development database:

```
$ rails generate migration add_picture_to_microposts picture:string
$ rails db:migrate
```