Adding image upload to the micropost create form.
On some systems, you may need to restart the Rails server at this point
to keep the test suite green. (If you’re using Guard as described in
Section 3.6.2: Automated tests with Guard, you may need to restart that
as well, and it may even be necessary to exit the terminal shell and
re-run Guard in a new one.)

To include the uploader on the Home page as in [Figure
13.18](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/micropost_image_mockup.png),
we need to include a file_field tag in the micropost form, as shown in
this commit.