Adding image display to microposts.
Once the image has been uploaded, we can render it using the image_tag
helper in the micropost partial, as shown in this commit. Notice the use
of the picture? boolean method to prevent displaying an image tag when
there isn’t an image. This method is created automatically by
CarrierWave based on the name of the image attribute. The result of
making a successful submission by hand appears in [Figure
13.20](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/microposts_with_image_4th_ed.png).
Writing an automated test for image upload is left as an exercise.

Exercises -

1. Upload a micropost with attached image. Does the result look too big?
(If so, don’t worry; we’ll fix it in Section 13.4.3).

Yes, the image looks too big.

Micropost images

2. Following the template in that is will be commented out for
```
test/integration/microposts_interface_test.rb
```
, write a test of the image uploader in Section 13.4. As preparation,
you should add an image to the fixtures directory (using, e.g, cp
app/assets/images/rails.png test/fixtures/). The additional assertions
in included in this commit, will check both for a file upload field on
the Home page and for a valid image attribute on the micropost resulting
from valid submission. Note the use of the special fixture_file_upload
method for uploading files as fixtures inside tests. Hint: To check for
a valid picture attribute, use the assigns method mentioned in Section
11.3.3: Activation test and refactoring to access the micropost in the
create action after valid submission.