The picture format validation.
The uploader in Section 13.4.1: Basic image upload is a good start, but
it has significant limitations. In particular, it doesn’t enforce any
constraints on the uploaded file, which can cause problems if users try
to upload large files or invalid file types. To remedy this defect,
we’ll add validations for the image size and format, both on the server
and on the client (i.e., in the browser).

The first image validation, which restricts uploads to valid image
types, appears in the CarrierWave uploader itself. The resulting code
(which appears as a commented-out suggestion in the generated uploader)
verifies that the image filename ends with a valid image extension (PNG,
GIF, and both variants of JPEG), as shown in this commit.