Adding validations to images.
The second validation, which controls the size of the image, appears in the Micropost model itself. In contrast to previous model validations, file size validation doesn’t correspond to a built-in Rails validator. As a result, validating images requires defining a custom validation, which we’ll call picture_size and define as shown in this commit. Note the use of validate (as opposed to validates) to call a custom validation.

This custom validation arranges to call the method corresponding to the given symbol (:picture_size). In picture_size itself, we add a custom error message to the errors collection (seen before briefly in Section 6.2.2: Validating presence ), in this case a limit of 5 megabytes (using a syntax seen before shown below).

Rails adds other helpers, too:

  >> 1.kilobyte
  => 1024
  >> 5.megabytes
  => 5242880
These are useful for upload validations, making it easy to restrict, say, image uploads to 5.megabytes.