Checking the file size with JQuery.
To go along with the validations in commit: The picture format validation and in this commit, we’ll add two client-side checks on the uploaded image. We’ll first mirror the format validation by using the accept parameter in the file_field input tag:

```
<%= f.file_field :picture, accept: 'image/jpeg,image/gif,image/png' %>
```

The valid formats consist of the MIME types accepted by the validation
in this commit: The picture format validation.

```
<%= f.file_field :picture, accept: 'image/jpeg,image/gif,image/png' %>
```

The valid formats consist of the MIME types accepted by the validation
in commit: The picture format validation.

Next, we’ll include a little JavaScript (or, more specifically, jQuery)
to issue an alert if a user tries to upload an image that’s too big
(which prevents accidental time-consuming uploads and lightens the load
on the server):

```
$('#micropost_picture').bind('change', function() {
  var size_in_megabytes = this.files[0].size/1024/1024;
  if (size_in_megabytes > 5) {
    alert('Maximum file size is 5MB. Please choose a smaller file.');
  }
});
```

Although jQuery isn’t the focus of this book, you might be able to
figure out that the code above monitors the page element containing the
CSS id micropost_picture (as indicated by the hash mark #), which is the
id of the micropost form in Listing 13.60. (The way to figure this out
is to Ctrl-click and use your browser’s web inspector.) When the element
with that CSS id changes, the jQuery function fires and issues the alert
method if the file is too big.19

The result of adding these additional checks appears in this commit.

As you can see by trying to upload a file that’s too big, the code in
this commit doesn’t actually clear the file input field, so a user can
just dismiss the alert box and continue trying to upload the file. If
this were a book on jQuery, we would probably polish this slight blemish
by clearing the field, but it’s important to understand that front-end
code like that shown in commit: Checking the file size with jQuery can’t
prevent a user from trying to upload a file that’s too big. Even if our
code cleared the file input field, there would be no way to prevent a
user from editing the JavaScript with a web inspector or issue a direct
POST request using, e.g., curl. To prevent users from uploading
arbitrarily large files, it is thus essential to include a server-side
validation, as in commit: Adding validations to images.

Exercises -

1. What happens if you try uploading an image bigger than 5 megabytes?

Using the
[sample-videos](http://www.sample-videos.com/download-sample-jpg-image.php),
trying to upload 10MB file.

A popup is displayed, display the following message: Maximum file size
is 5MB. Please choose a smaller file. in the browser.

2. What happens if you try uploading a file with an invalid extension?

I was able to upload a file that was a png went ahead and changed the
extension to php and it was able to be upload with no errors.