Configuring the image uploader for image resizing.
The image size validations in Section 13.4.2: Image validation are a
good start, but they still allow the uploading of images large enough to
break our site’s layout, sometimes with frightening results (Figure
13.21). Thus, while it’s convenient to allow users to select fairly
large images from their local disk, it’s also a good idea to resize the
images before displaying them.

We’ll be resizing images using the image manipulation program
ImageMagick, which we need to install on the development environment.
(As we’ll see in Section 13.4.4: Image upload in production, when using
Heroku for deployment ImageMagick comes pre-installed in production.) On
the cloud IDE, we can do this as follows:

```
$ sudo apt-get update
$ sudo apt-get install imagemagick --fix-missing
```

(If you’re developing on your local machine, you may have to install
ImageMagick a different way, such as using Homebrew on a Mac via brew
install imagemagick. Use your technical sophistication (Box 1.1) if you
get stuck.)

Next, we need to include CarrierWave’s MiniMagick interface for
ImageMagick, together with a resizing command. For the resizing command,
there are several possibilities listed in the MiniMagick documentation,
but the one we want is resize_to_limit: [400, 400], which resizes large
images so that they aren’t any bigger than 400px in either dimension,
while simultaneously leaving smaller images alone. (The other main
possibilities listed in the CarrierWave documentation on MiniMagick
stretch images if they’re too small, which isn’t what we want in this
case.) With the code as shown in this commit, large images are now
resized nicely [Figure
13.22](https://www.railstutorial.org/book/user_microposts#fig-resized_image).