Configuring the image uploader for production.
The image uploader developed in Section 13.4.3:Image resizing is good enough for development, but (as seen in the storage :file line in commit: Configuring the image uploader for image resizing.) it uses the local filesystem for storing the images, which isn’t a good practice in production. Instead, we’ll use a cloud storage service to store images separately from our application.

To configure our application to use cloud storage in production, we’ll use the fog gem, as shown in this commit.

There are many choices for cloud storage, but we’ll use one of the most popular and well-supported, Amazon.com’s Simple Storage Service (S3). Here are the essential steps to getting set up:

1. Sign up for an Amazon Web Services account if you don’t have one already. If you signed up for the Cloud9 IDE in Section 1.2.1: Development environment, you already have an AWS account and can skip this step.

2. Create a user via AWS Identity and Access Management (IAM) and record the access key and secret key.

3. Create an S3 bucket (with a name of your choice) using the AWS Console, and then grant read and write permission to the user created in the previous step.