Implementing the active relationships has_many association.
Before implementing user following and followers, we first need to
establish the association between users and relationships. A user
has_many relationships, and—since relationships involve two users—a
relationship belongs_to both a follower and a followed user.

As with microposts in Section 13.1.3: User/Micropost associations, we
will create new relationships using the user association, with code such
as

```
user.active_relationships.build(followed_id: ...)
```
At this point, you might expect application code as in Section 13.1.3:
User/Micropost associations, and it’s similar, but there are two key
differences.

First, in the case of the user/micropost association we could write

```
class User < ApplicationRecord
  has_many :microposts
  .
  .
  .
end
```

This works because by convention Rails looks for a Micropost model
corresponding to the :microposts symbol. In the present case, though, we
want to write

```
has_many :active_relationships
```

even though the underlying model is called Relationship. We will thus
have to tell Rails the model class name to look for.

Second, before we wrote

```
class Micropost < ApplicationRecord
  belongs_to :user
  .
  .
  .
end
```
in the Micropost model. This works because the microposts table has a
user_id attribute to identify the user (Section 13.1.1: The basic model
). An id used in this manner to connect two database tables is known as
a foreign key, and when the foreign key for a User model object is
user_id, Rails infers the association automatically: by default, Rails
expects a foreign key of the form <class>_id, where <class> is the
lower-case version of the class name. In the present case, although we
are still dealing with users, the user following another user is now
identified with the foreign key follower_id, so we have to tell that to
Rails.

The result of the above discussion is the user/relationship association
shown in this commit.