Adding the User model following association.
Followed users

We come now to the heart of the Relationship associations: following and
followers. Here we will use has_many :through for the first time: a user
has many following through relationships, as illustrated in commit: . By
default, in a has_many :through association Rails looks for a foreign
key corresponding to the singular version of the association. In other
words, with code like

```
has_many :followeds, through: :active_relationships
```

Rails would see “followeds” and use the singular “followed”, assembling
a collection using the followed_id in the relationships table. But, as
noted in Section 14.1.1: A problem with the data model (and a solution),
user.followeds is rather awkward, so we’ll write user.following instead.
Naturally, Rails allows us to override the default, in this case using
the source parameter (as shown in this commit), which explicitly tells
Rails that the source of the following array is the set of followed ids.