Tests for some “following” utility methods. red
The association defined in commit: Adding the User model following
association leads to a powerful combination of Active Record and
array-like behavior. For example, we can check if the followed users
collection includes another user with the include? method (Section
4.3.1: Arrays and ranges), or find objects through the association:

```
user.following.include?(other_user)
user.following.find(other_user)
```
We can also add and delete elements just as with arrays:

```
user.following << other_user
user.following.delete(other_user)
```

(Recall from Section 4.3.1 that the shovel operator << appends to the
end of an array.)

Although in many contexts we can effectively treat following as an
array, Rails is smart about how it handles things under the hood. For
example, code like

```
following.include?(other_user)
```
looks like it might have to pull all the followed users out of the
database to apply the include? method, but in fact for efficiency Rails
arranges for the comparison to happen directly in the database. (Compare
to the code in Section 13.2.1: Rendering microposts, where we saw that

```
user.microposts.count
```
performs the count directly in the database.)

To manipulate following relationships, we’ll introduce follow and
unfollow utility methods so that we can write, e.g.,
user.follow(other_user). We’ll also add an associated following? boolean
method to test if one user is following another.

This is exactly the kind of situation where I like to write some tests
first. The reason is that we are quite far from writing a working web
interface for following users, but it’s hard to proceed without some
sort of client for the code we’re developing. In this case, it’s easy to
write a short test for the User model, in which we use following? to
make sure the user isn’t following the other user, use follow to follow
another user, use following? to verify that the operation succeeded, and
finally unfollow and verify that it worked. The result appears in this
commit.

```
$ rails tests
```

ERROR["test_should_follow_and_unfollow_a_user", UserTest,
23.63095272798091]
 test_should_follow_and_unfollow_a_user#UserTest (23.63s)
NoMethodError:         NoMethodError: undefined method `following?' for
#<User:0x00000013179ce8>
            test/models/user_test.rb:91:in `block in <class:UserTest>'

E