Implementing user.followers using passive relationships.
The final piece of the relationships puzzle is to add a user.followers
method to go with user.following. You may have noticed from Figure 14.7
that all the information needed to extract an array of followers is
already present in the relationships table (which we are treating as the
active_relationships table via the code in commit: Implementing the
active relationships has_many association). Indeed, the technique is
exactly the same as for followed users, with the roles of follower_id
and followed_id reversed, and with passive_relationships in place of
active_relationships. The data model then appears as in Figure 14.9.

Figure 14.9: A model for user followers through passive relationships.

     |                user                      |
     |------------------------------------------|
     |  id |   name        | email              |
     |------------------------------------------|
|----|   1 | Michael Hartl | mhart1@example.com |
|
|
|            |   passive_relationships   |           | user.followers  |
|            | follower_id | followed_id |           |id|name | email  |
|  through   |---------------------------|           |-----------------|
|----------->| 2           | 1           |---------->| 3|     | ...    |
|----------->| 7           | 1           |---------->| 2|     | ...    |
|----------->| 1           | 3           |    |----->| 9|     | ...    |
|----------->| 2           | 7           |    |
|----------->| 10          | 1           |    |
|----------->| 1           | 2           |----|
|----------->| 8           | 1           |
|----------->| 1           | 9           |

It’s worth noting that we could actually omit the :source key for
followers in this commit, using simply

has_many :followers, through: :passive_relationships
This is because, in the case of a :followers attribute, Rails will
singularize “followers” and automatically look for the foreign key
follower_id in this case. In this commit, keeps the :source key to
emphasize the parallel structure with the has_many :following
association.