Adding following/follower relationships to the sample data.
Section 14.1: The Relationship model placed some rather heavy demands on
our data modeling skills, and it’s fine if it takes a while to soak in.
In fact, one of the best ways to understand the associations is to use
them in the web interface.

In the introduction to this chapter, we saw a preview of the page flow
for user following. In this section, we will implement the basic
interface and following/unfollowing functionality shown in those
mockups. We will also make separate pages to show the user following and
followers arrays. In Section 14.3: The status feed, we’ll complete our
sample application by adding the user’s status feed.

Sample following data

As in previous chapters, we will find it convenient to use rails db:seed
to fill the database with sample relationships. This will allow us to
design the look and feel of the web pages first, deferring the back-end
functionality until later in this section.

Code to seed the following relationships appears in Listing 14.14. Here
we somewhat arbitrarily arrange for the first user to follow users 3
through 51, and then have users 4 through 41 follow that user back. The
resulting relationships will be sufficient for developing the
application interface.

To execute the code in Listing 14.14, we reset and reseed the database
as usual:

```
$ rails db:migrate:reset
$ rails db:seed
```
Exercises -

1. Using the console, confirm that User.first.followers.count matches
the value expected from this commit.
 In this commit, we've created followers with a range from 3 - 40. That
equals to 38 followers.

 After running User.first.followers.count, it returns:

 >> User.first.followers.count
User.first.followers.count
  User Load (0.0ms)  SELECT  "users".* FROM "users" ORDER BY
"users"."id" ASC LIMIT ?  [["LIMIT", 1]]
   (0.0ms)  SELECT COUNT(*) FROM "users" INNER JOIN "relationships" ON
"users"."id" = "relationships"."follower_id" WHERE
"relationships"."followed_id" = ?  [["followed_id", 1]]
=> 38

2. Confirm that User.first.following.count is correct as well.

In this commit, we've created following with a range from 2 - 50. That
equals to 49 followers.

>> User.first.following.count
User.first.following.count
  User Load (0.0ms)  SELECT  "users".* FROM "users" ORDER BY
"users"."id" ASC LIMIT ?  [["LIMIT", 1]]
   (0.0ms)  SELECT COUNT(*) FROM "users" INNER JOIN "relationships" ON
"users"."id" = "relationships"."followed_id" WHERE
"relationships"."follower_id" = ?  [["follower_id", 1]]
=> 49