Adding following and followers actions to the Users controller.
Now that our sample users have both followed users and followers, we
need to update the profile page and Home page to reflect this. We’ll
start by making a partial to display the following and follower
statistics on the profile and home pages. We’ll next add a
follow/unfollow form, and then make dedicated pages for showing
“following” (followed users) and “followers”.

As noted in Section 14.1.1: A problem with the data model (and a
solution), we’ll adopt Twitter’s convention of using “following” as a
label for followed users, as in “50 following”. This usage is reflected
in the mockup sequence starting in [Figure
14.1](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/page_flow_profile_mockup_3rd_edition.png)
and shown in close-up in [Figure
14.10](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/stats_partial_mockup.png).

The stats in [Figure
14.10](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/stats_partial_mockup.png)
consist of the number of users the current user is following and the
number of followers, each of which should be a link to its respective
dedicated display page. In Chapter 5, we stubbed out such links with the
dummy text ’#’, but that was before we had much experience with routes.
This time, although we’ll defer the actual pages to Section 14.2.3:
Following and followers pages, we’ll make the routes now, as seen in
this commit. This code uses the :member method inside a resources block,
which we haven’t seen before, but see if you can guess what it does.

You might suspect that the URLs for following and followers will look
like /users/1/following and /users/1/followers, and that is exactly what
the code in this commit arranges. Since both pages will be showing data,
the proper HTTP verb is a GET request, so we use the get method to
arrange for the URLs to respond appropriately. Meanwhile, the member
method arranges for the routes to respond to URLs containing the user
id. The other possibility, collection, works without the id, so that

```
resources :users do
  collection do
    get :tigers
  end
end
```

would respond to the URL /users/tigers (presumably to display all the
tigers in our application).

A table of the routes generated by this commit appears in Table 14.2.
Note the named routes for the followed user and followers pages, which
we’ll put to use shortly.