A partial for displaying follower stats.
With the routes defined, we are now in a position to define the stats
partial, which involves a couple of links inside a div, as shown in this
commit.

Since we will be including the stats on both the user show pages and the
home page, the first line of in this commit picks the right one using

```
<% @user ||= current_user %>
```

As discussed in Box 8.1, this does nothing when @user is not nil (as on
a profile page), but when it is (as on the Home page) it sets @user to
the current user. Note also that the following/follower counts are
calculated through the associations using

```
@user.following.count
```
and

```
@user.followers.count
```
Compare these to the microposts count from in commit: Adding microposts
to the user show page, where we wrote

```
@user.microposts.count
```

to count the microposts. As in that case, Rails calculates the count
directly in the database for efficiency.

One final detail worth noting is the presence of CSS ids on some
elements, as in

```
<strong id="following" class="stat">
...
</strong>
```

This is for the benefit of the Ajax implementation in Section 14.2.5: A
working follow button with Ajax, which accesses elements on the page
using their unique ids.