A form for following/unfollowing a user.
Listing 14.21: A form for following a user.
app/views/users/_follow.html.erb

Listing 14.22: A form for unfollowing a user.
app/views/users/_unfollow.html.erb

These two forms both use form_for to manipulate a Relationship model
object; the main difference between the two is that Listing 14.21 builds
a new relationship, whereas Listing 14.22 finds the existing
relationship. Naturally, the former sends a POST request to the
Relationships controller to create a relationship, while the latter
sends a DELETE request to destroy a relationship. (We’ll write these
actions in Section 14.2.4: A working follow button the standard way)
Finally, you’ll note that the follow form doesn’t have any content other
than the button, but it still needs to send the followed_id to the
controller. We accomplish this with the hidden_field_tag method in
Listing 14.21, which produces HTML of the form

```
<input id="followed_id" name="followed_id" type="hidden" value="3" />
```

As we saw in Section 12.3: Resetting the password (in commit: The form
to reset a password), the hidden input tag puts the relevant information
on the page without displaying it in the browser.