Adding the follow form and follower stats to the user profile page.
We can now include the follow form and the following statistics on the
user profile page simply by rendering the partials, as shown in this
commit. Profiles with follow and unfollow buttons, respectively, appear
in [Figure
14.12](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/profile_follow_button_3rd_edition.png)
and [Figure
14.13](https://www.railstutorial.org/book/following_users#fig-profile_unfollow_button).

We’ll get these buttons working soon enough—in fact, we’ll do it two
ways, the standard way (Section 14.2.4: A working follow button the
standard way) and using Ajax (Section 14.2.5: A working follow button
with Ajax)—but first we’ll finish the HTML interface by making the
following and followers pages.

1. Verify that /users/2 has a follow form and that /users/5 has an
unfollow form. Is there a follow form on /users/1?

After login as Example User(user: example@railstutorial.org,
pasword:forbar), then I was able to see that /users/2 has a follow form
and that /users/5 has an unfollow form. No, I do not have follow form on
/users/1.

2. Confirm in the browser that the stats appear correctly on the Home
page and on the profile page.

Yes, the stats appear correctly on the Home page and on the profile
page.

3. Write tests for the stats on the Home page. Hint: Add to the test in
commit: A test for the user profile. Why don’t we also have to test the
stats on the profile page?

We don’t have to test the stats on the profile page because they are the
same page.  user/show.html.erb