Tests for the authorization of the following and followers pages. red
Following and followers pages

Pages to display followed users and followers will resemble a hybrid of
the user profile page and the user index page (Section 10.3.1: Users
index), with a sidebar of user information (including the following
stats) and a list of users. In addition, we’ll include a raster of
smaller user profile image links in the sidebar. Mockups matching these
requirements appear in [Figure
14.14](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/following_mockup_bootstrap.png)
(following) and [Figure
14.15](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/followers_mockup_bootstrap.png)
(followers).

Our first step is to get the following and followers links to work.
We’ll follow Twitter’s lead and have both pages require user login. As
with most previous examples of access control, we’ll write the tests
first, as shown in commit: Tests for the authorization of the following
and followers pages. Note that commit: Tests for the authorization of
the following and followers pages uses the named routes from Table 14.2.